PWGCF: FemtoUniverse -- fixed nSigmaPIDMax from int to float and changed pT to p for PID (#4246)
* fixed nSigmaPIDMax from int to float

* particle identification from pT to p

---------

+++ PWGCF/FemtoUniverse/Core/FemtoUniverseTrackSelection.h
@@ -197,7 +197,7 @@ class FemtoUniverseTrackSelection : public FemtoUniverseObjectSelection<float, f
     return outString;
   }
 
-  int getSigmaPIDMax()
+  float getSigmaPIDMax()
   {
     return nSigmaPIDMax;
   }
